Rename: 페이지 레이아웃 컴포넌트 폴더로 이동
