+++ src/components/feature/index.ts
@@ -1,2 +1,3 @@
 export * from './group'
 export * from './invite'
+export * from './layout'
