+++ src/components/feature/layout/group-member/index.ts
@@ -1,0 +1,1 @@
+export { GroupMemberLayout } from './GroupMemberLayout'
