+++ src/components/feature/layout/main/error-fallback/index.ts
@@ -1,0 +1,1 @@
+export { GlobalErrorFallback } from './ErrorFallback'
