+++ src/components/feature/layout/main/group-section/create-group-button/index.ts
@@ -1,0 +1,1 @@
+export { CreateGroupButton } from './CreateGroupButton'
