+++ src/components/feature/layout/main/group-section/error-fallback/index.ts
@@ -1,0 +1,1 @@
+export { GroupErrorFallback } from './GroupErrorFallback'
