+++ src/components/feature/layout/main/group-section/group-list/index.ts
@@ -1,0 +1,1 @@
+export { GroupList } from './GroupList'
