+++ src/components/feature/layout/main/group-section/index.ts
@@ -1,0 +1,1 @@
+export { GroupSection, GroupSectionSkeleton } from './GroupSection'
