+++ src/components/feature/layout/main/member-section/error-fallback/index.ts
@@ -1,0 +1,1 @@
+export { MemberErrorFallback } from './ErrorFallback'
