+++ src/components/feature/layout/main/member-section/friend-section/header/index.ts
@@ -1,0 +1,1 @@
+export { FriendHeader } from './Header'
