+++ src/components/feature/layout/main/member-section/friend-section/index.ts
@@ -1,0 +1,1 @@
+export { FriendSection, FriendSectionSkeleton } from './FriendSection'
