+++ src/components/feature/layout/main/member-section/index.ts
@@ -1,0 +1,1 @@
+export { MemberSection } from './MemberSection'
