+++ src/components/feature/layout/main/member-section/select-friend-section/member-list/index.ts
@@ -1,0 +1,1 @@
+export { MemberList } from './MemberList'
