+++ src/components/feature/layout/profile-question/index.ts
@@ -1,0 +1,1 @@
+export { ProfileQuestionLayout } from './ProfileQuestionLayout'
