+++ src/components/feature/layout/profile-question/question-section/create-question-button/index.ts
@@ -1,0 +1,1 @@
+export { CreateQuestionButton } from './CreateQuestionButton'
