+++ src/components/feature/layout/profile-question/question-section/index.ts
@@ -1,0 +1,1 @@
+export { QuestionSection } from './QuestionSection'
