+++ src/components/feature/layout/profile-question/question-section/question-list/index.ts
@@ -1,0 +1,1 @@
+export { QuestionListSection } from './QuestionList'
